FIX: Make sure roles are always returned as list
WHY the array_values:
The way doctrine saves these roles inside the database sometimes differs (no idea why.).
Sometimes doctrine will save them as array list and sometimes it saves them as associative array.
However we expect the structure to always be a list, which is why we make sure here, that we actually get one ;)