FIX: Crash on empty description field
The CKEeditor currently seems to be bugged and does not respect the
'required' constraint at all. Therefore we had to fallback to symfonys
own validation which is done by setting constraint annotations to the
entity. However that also means, that the property in question has to
have a setter with an optional parameter because the form is processed,
sets the entities and only then can be validated.

If the setter does not allow for null values, the app will crash before
we can even try to validate the form.